// src/common/errors/error-codes.enum.ts
export enum ErrorCode {
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  BAD_REQUEST = 'BAD_REQUEST',
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  EMAIL_ALREADY_USED = 'EMAIL_ALREADY_USED',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  INVALID_OTP = 'INVALID_OTP',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  CACHE_ATTEMPT_FAILED = 'CACHE_ATTEMPT_FAILED',
  GAME_NOT_FOUND = 'GAME_NOT_FOUND',
  GAME_INVALID = 'GAME_INVALID',
  NOT_YOUR_GAME = 'NOT_YOUR_GAME',
  INVALID_MOVE = 'INVALID_MOVE',
  NOT_YOUR_TURN = 'NOT_YOUR_TURN',
  NO_MOVES_TO_UNDO = 'NO_MOVES_TO_UNDO',
  UNAUTHORIZED_MOVE = 'UNAUTHORIZED_MOVE',
  UNDO_NOT_ALLOWED = 'UNDO_NOT_ALLOWED',
}